Cat bug fix

+++ models/categories.go
@@ -3,7 +3,7 @@ package models
 import "database/sql"
 
 type Category struct {
-	ID       int64         `db:"id" json:"id"`
+	ID       int           `db:"id" json:"id"`
 	Name     string        `db:"name" json:"name"`
 	ParentID sql.NullInt64 `db:"parent_id" json:"parentId"`
 }
