+++ models/channel.go
@@ -5,14 +5,14 @@ import (
 )
 
 type Channel struct {
-	ID          int64          `db:"id" json:"id"`
+	ID          int            `db:"id" json:"id"`
 	Title       string         `db:"title" json:"title"`
 	Description string         `db:"description" json:"description"`
 	Keywords    sql.NullString `db:"keywords" json:"-"`
 	Image       string         `db:"image" json:"image"`
 	URL         string         `db:"url" json:"url"`
 	Website     sql.NullString `db:"website" json:"website"`
-	NumPodcasts int64          `db:"num_podcasts" json:"numPodcasts"`
+	NumPodcasts int            `db:"num_podcasts" json:"numPodcasts"`
 	Podcasts    []*Podcast     `db:"-" json:"-"`
 	Categories  []string       `db:"-" json:"-"`
 }
