+++ models/page.go
@@ -4,21 +4,21 @@ import "math"
 
 const defaultPageSize = 10
 
-func NewPaginator(page int64, numRows int64) *Paginator {
+func NewPaginator(page int, numRows int) *Paginator {
 	p := &Paginator{
 		CurrentPage: page,
 		NumRows:     numRows,
 		PageSize:    defaultPageSize,
 	}
-	p.NumPages = int64(math.Ceil(float64(numRows) / float64(defaultPageSize)))
+	p.NumPages = int(math.Ceil(float64(numRows) / float64(defaultPageSize)))
 	p.Offset = (page - 1) * defaultPageSize
 	return p
 }
 
 type Paginator struct {
-	NumRows     int64 `json:"numRows"`
-	NumPages    int64 `json:"numPages"`
-	CurrentPage int64 `json:"page"`
-	PageSize    int64 `json:"pageSize"`
-	Offset      int64 `json:"-"`
+	NumRows     int `json:"numRows"`
+	NumPages    int `json:"numPages"`
+	CurrentPage int `json:"page"`
+	PageSize    int `json:"pageSize"`
+	Offset      int `json:"-"`
 }
