+++ models/plays.go
@@ -3,6 +3,6 @@ package models
 import "time"
 
 type Play struct {
-	PodcastID int64     `db:"podcast_id" json:"podcastId"`
+	PodcastID int       `db:"podcast_id" json:"podcastId"`
 	CreatedAt time.Time `db:"created_at" json:"createdAt"`
 }
