+++ models/user.go
@@ -6,13 +6,13 @@ import (
 )
 
 type User struct {
-	ID            int64     `db:"id" json:"id"`
+	ID            int       `db:"id" json:"id"`
 	Name          string    `db:"name" json:"name"`
 	Email         string    `db:"email" json:"email"`
 	Password      string    `db:"password" json:"-"`
 	CreatedAt     time.Time `db:"created_at" json:"-"`
-	Bookmarks     []int64   `db:"-" json:"bookmarks"`
-	Subscriptions []int64   `db:"-" json:"subscriptions"`
+	Bookmarks     []int     `db:"-" json:"bookmarks"`
+	Subscriptions []int     `db:"-" json:"subscriptions"`
 	Plays         []Play    `db:"-" json:"plays"`
 }
