+++ store/plays.go
@@ -6,12 +6,12 @@ import (
 )
 
 type PlayReader interface {
-	SelectByUserID(DataHandler, *[]models.Play, int64) error
+	SelectByUserID(DataHandler, *[]models.Play, int) error
 }
 
 type PlayWriter interface {
-	Create(DataHandler, int64, int64) error
-	DeleteAll(DataHandler, int64) error
+	Create(DataHandler, int, int) error
+	DeleteAll(DataHandler, int) error
 }
 
 type PlayStore interface {
@@ -33,20 +33,20 @@ func newPlayStore() PlayStore {
 
 type playSqlReader struct{}
 
-func (r *playSqlReader) SelectByUserID(dh DataHandler, plays *[]models.Play, userID int64) error {
+func (r *playSqlReader) SelectByUserID(dh DataHandler, plays *[]models.Play, userID int) error {
 	q := "SELECT podcast_id, created_at FROM plays WHERE user_id=$1"
 	return sqlx.Select(dh, plays, q, userID)
 }
 
 type playSqlWriter struct{}
 
-func (w *playSqlWriter) Create(dh DataHandler, podcastID, userID int64) error {
+func (w *playSqlWriter) Create(dh DataHandler, podcastID, userID int) error {
 	q := "SELECT add_play($1, $2)"
 	_, err := dh.Exec(q, podcastID, userID)
 	return err
 }
 
-func (w *playSqlWriter) DeleteAll(dh DataHandler, userID int64) error {
+func (w *playSqlWriter) DeleteAll(dh DataHandler, userID int) error {
 	q := "DELETE FROM plays WHERE user_id=$1"
 	_, err := dh.Exec(q, userID)
 	return err
